+++ tests/testthat/test-from_json_conn.R
@@ -1,10 +1,10 @@
 
 test_that("parsing json from connection works", {
   
-  js1 <- from_json_file("examples/test.json")
+  js1 <- read_json_file("examples/test.json")
   
   con <- gzfile("examples/test.json.gz")
-  js2 <- from_json_conn(gzfile("examples/test.json.gz"))
+  js2 <- read_json_conn(gzfile("examples/test.json.gz"))
   close.connection(con)
   
   expect_identical(js1, js2)