+++ tests/testthat/test-handling-specials.R
@@ -5,17 +5,17 @@ test_that("to_json handling specials works", {
   # Special Logical
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    to_json_str(c(TRUE, NA)),
+    write_json_str(c(TRUE, NA)),
     '[true,null]'
   )
   
   expect_identical(
-    to_json_str(c(TRUE, NA), opts = to_opts(num_specials = 'string')),
+    write_json_str(c(TRUE, NA), opts = to_opts(num_specials = 'string')),
     '[true,"NA"]'
   )
   
   expect_identical(
-    to_json_str(c(TRUE, NA), opts = to_opts(num_specials = 'null')),
+    write_json_str(c(TRUE, NA), opts = to_opts(num_specials = 'null')),
     '[true,null]'
   )
   
@@ -24,17 +24,17 @@ test_that("to_json handling specials works", {
   # Special INTs
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    to_json_str(c(0L, NA_integer_)),
+    write_json_str(c(0L, NA_integer_)),
     '[0,null]'
   )
   
   expect_identical(
-    to_json_str(c(0L, NA_integer_), opts = to_opts(num_specials = 'string')),
+    write_json_str(c(0L, NA_integer_), opts = to_opts(num_specials = 'string')),
     '[0,"NA"]'
   )
   
   expect_identical(
-    to_json_str(c(0L, NA_integer_), opts = to_opts(num_specials = 'null')),
+    write_json_str(c(0L, NA_integer_), opts = to_opts(num_specials = 'null')),
     '[0,null]'
   )
   
@@ -43,17 +43,17 @@ test_that("to_json handling specials works", {
   # Special STRING
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    to_json_str(c("a", NA_character_)),
+    write_json_str(c("a", NA_character_)),
     '["a",null]'
   )
   
   expect_identical(
-    to_json_str(c("a", NA_character_), opts = to_opts(str_specials = 'string')),
+    write_json_str(c("a", NA_character_), opts = to_opts(str_specials = 'string')),
     '["a","NA"]'
   )
   
   expect_identical(
-    to_json_str(c("a", NA_character_), opts = to_opts(str_specials = 'null')),
+    write_json_str(c("a", NA_character_), opts = to_opts(str_specials = 'null')),
     '["a",null]'
   )
   
@@ -62,17 +62,17 @@ test_that("to_json handling specials works", {
   # Special Numeric
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    to_json_str(c(0, NA_real_, Inf, -Inf, NaN)),
+    write_json_str(c(0, NA_real_, Inf, -Inf, NaN)),
     '[0.0,null,null,null,null]'
   )
   
   expect_identical(
-    to_json_str(c(0, NA_real_, Inf, -Inf, NaN), opts = to_opts(num_specials = 'string')),
+    write_json_str(c(0, NA_real_, Inf, -Inf, NaN), opts = to_opts(num_specials = 'string')),
     '[0.0,"NA","Inf","-Inf","NaN"]'
   )
   
   expect_identical(
-    to_json_str(c(0, NA_real_, Inf, -Inf, NaN), opts = to_opts(num_specials = 'null')),
+    write_json_str(c(0, NA_real_, Inf, -Inf, NaN), opts = to_opts(num_specials = 'null')),
     '[0.0,null,null,null,null]'
   )
 })
@@ -84,17 +84,17 @@ test_that("from_json handling specials works", {
   # Logical
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    from_json_str('[true,"NA"]'),
+    read_json_str('[true,"NA"]'),
     c(TRUE, NA)
   )  
   
   expect_identical(
-    from_json_str('[true,"NA"]', opts = from_opts(num_specials = "special")),
+    read_json_str('[true,"NA"]', opts = from_opts(num_specials = "special")),
     c(TRUE, NA)
   )  
   
   expect_identical(
-    from_json_str('[true,"NA"]', opts = from_opts(num_specials = "string")),
+    read_json_str('[true,"NA"]', opts = from_opts(num_specials = "string")),
     list(TRUE, "NA")
   )  
   
@@ -102,17 +102,17 @@ test_that("from_json handling specials works", {
   # Integer
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    from_json_str('[0, "NA"]'),
+    read_json_str('[0, "NA"]'),
     c(0L, NA_integer_)
   )  
   
   expect_identical(
-    from_json_str('[0, "NA"]', opts = from_opts(num_specials = "special")),
+    read_json_str('[0, "NA"]', opts = from_opts(num_specials = "special")),
     c(0L, NA_integer_)
   )  
   
   expect_identical(
-    from_json_str('[0, "NA"]', opts = from_opts(num_specials = "string")),
+    read_json_str('[0, "NA"]', opts = from_opts(num_specials = "string")),
     list(0L, "NA")
   )  
   
@@ -121,17 +121,17 @@ test_that("from_json handling specials works", {
   # Numeric
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    from_json_str('[0.0, "NA", "Inf", "-Inf", "NaN"]'),
+    read_json_str('[0.0, "NA", "Inf", "-Inf", "NaN"]'),
     c(0, NA_real_, Inf, -Inf, NaN)
   )  
   
   expect_identical(
-    from_json_str('[0.0, "NA", "Inf", "-Inf", "NaN"]', opts = from_opts(num_specials = "special")),
+    read_json_str('[0.0, "NA", "Inf", "-Inf", "NaN"]', opts = from_opts(num_specials = "special")),
     c(0, NA_real_, Inf, -Inf, NaN)
   )  
   
   expect_identical(
-    from_json_str('[0.0, "NA", "Inf", "-Inf", "NaN"]', opts = from_opts(num_specials = "string")),
+    read_json_str('[0.0, "NA", "Inf", "-Inf", "NaN"]', opts = from_opts(num_specials = "string")),
     list(0, "NA", "Inf", "-Inf", "NaN")
   )  
   
@@ -139,17 +139,17 @@ test_that("from_json handling specials works", {
   # String
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   expect_identical(
-    from_json_str('["a", "NA"]'),
+    read_json_str('["a", "NA"]'),
     c("a", "NA")
   )
   
   expect_identical(
-    from_json_str('["a", "NA"]', opts = from_opts(str_specials = "string")),
+    read_json_str('["a", "NA"]', opts = from_opts(str_specials = "string")),
     c("a", "NA")
   )
   
   expect_identical(
-    from_json_str('["a", "NA"]', opts = from_opts(str_specials = "special")),
+    read_json_str('["a", "NA"]', opts = from_opts(str_specials = "special")),
     c("a", NA_character_)
   )