+++ tests/testthat/test-json_promotion.R
@@ -10,12 +10,12 @@ js <- r"(
 
 test_that("numeric promotion to string instead of list works", {
 
-  res <- from_json_str(js)
+  res <- read_json_str(js)
   expect_identical(res$a, list(1.234560, 2L, "apple"))
   
-  res <- from_json_str(js, promote_num_to_string = FALSE)
+  res <- read_json_str(js, promote_num_to_string = FALSE)
   expect_identical(res$a, list(1.234560, 2L, "apple"))
 
-  res <- from_json_str(js, promote_num_to_string = TRUE) 
+  res <- read_json_str(js, promote_num_to_string = TRUE) 
   expect_identical(res$a, c("1.234560", "2", "apple"))
 })