+++ tests/testthat/test-yyjson-flags.R
@@ -3,12 +3,12 @@ test_that("yyjson flags work", {
 
   expect_error(
     capture_output(
-      from_json_str('[1,2,3,]'), 'comma'
+      read_json_str('[1,2,3,]'), 'comma'
     )
   )
   
   expect_equal(
-    from_json_str('[1,2,3,]', yyjson_read_flag = read_flag$YYJSON_READ_ALLOW_TRAILING_COMMAS),
+    read_json_str('[1,2,3,]', yyjson_read_flag = read_flag$YYJSON_READ_ALLOW_TRAILING_COMMAS),
     c(1L, 2L, 3L)
   )