+++ NEWS.md
@@ -32,7 +32,7 @@
 # yyjsonr 0.1.1  2023-08-19
 
 * Update to option setting to remove overhead
-* Added `from_json_raw()` to parse JSON directly from a raw vector (which
+* Added `read_json_raw()` to parse JSON directly from a raw vector (which
   contains UTF-8 string data)
 
 # yyjsonr 0.1.0  2023-08-16